[util/math] add some basic operations of vec3

+++ util/ray_def.h
@@ -8,7 +8,7 @@
 
 namespace raycer {
 
-typedef double
+typedef float
     fnum;  // change the definition of num to gain higher precision, but
            // this is not necessary to use double since float is enough
